move config to utils
